backend/requirements.txt
Flask==2.0.3
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
elasticsearch==7.17.9
python-dotenv==0.21.0
Werkzeug==2.0.3
Pillow==9.5.0
markdown==3.3.7
SQLAlchemy==1.4.49 